+++ CHANGELOG.md
@@ -154,6 +154,7 @@
 - Dev: Added signal to invalidate paint buffers of channel views without forcing a relayout. (#5123)
 - Dev: Specialize `Atomic<std::shared_ptr<T>>` if underlying standard library supports it. (#5133)
 - Dev: Added the `developer_name` field to the Linux AppData specification. (#5138)
+- Dev: Added estimation for image sizes to avoid layout shifts. (#5192)
 
 ## 2.4.6